requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

annotated-types==0.6.0
anyio==4.3.0
atproto==0.0.46
azure-functions==1.18.0
blurhash==1.1.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
dnspython==2.6.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
libipld==1.2.3
Mastodon.py==1.8.1
oauthlib==3.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
sniffio==1.3.1
tweepy==4.14.0
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
